Refactor task management for multi-user collaboration and user-specific reminders
*   Implement participant-specific reminders by adding a `reminders` column to the `TaskParticipantsTable` and updating server routes to handle user-level reminder overrides.
*   Enhance task security and permissions by allowing room members (in addition to the task creator) to update and delete shared tasks.
*   Integrate FCM push notifications to alert users when they are added to a new task.
*   Optimize room member profile fetching in `TaskEditorViewModel` by using parallel coroutines (`async`/`awaitAll`) for improved performance.
*   Update the UI in `TaskEditorScreen` and `TaskTimelineComponent` to display participant names and grouped avatars instead of just counts.
*   Fix type-safety issues in `RuleBasedNlpEngine` by migrating day-difference calculations to use `Long` literals.
*   Include utility Python scripts to automate task route updates and ViewModel refactoring.

+++ core/src/commonMain/kotlin/com/yusufteker/pulse/core/ai/RuleBasedNlpEngine.kt
@@ -723,17 +723,17 @@ class RuleBasedNlpEngine {
         val targetLocal = kotlinx.datetime.Instant.fromEpochMilliseconds(epochMs).toLocalDateTime(timeZone)
         val currentLocal = kotlinx.datetime.Instant.fromEpochMilliseconds(getCurrentTimeMs()).toLocalDateTime(timeZone)
         
-        val diffDays = targetLocal.date.toEpochDays() - currentLocal.date.toEpochDays()
+        val diffDays = (targetLocal.date.toEpochDays() - currentLocal.date.toEpochDays()).toLong()
 
         val hour = targetLocal.hour
         val minute = targetLocal.minute
         val timeStr = "${hour.toString().padStart(2, '0')}:${minute.toString().padStart(2, '0')}"
 
         return when {
-            diffDays < 0 -> "geçmiş bir zaman"
-            diffDays == 0 -> "bugün saat $timeStr"
-            diffDays == 1 -> "yarın saat $timeStr"
-            diffDays < 7 -> "$diffDays gün sonra, saat $timeStr"
+            diffDays < 0L -> "geçmiş bir zaman"
+            diffDays == 0L -> "bugün saat $timeStr"
+            diffDays == 1L -> "yarın saat $timeStr"
+            diffDays < 7L -> "$diffDays gün sonra, saat $timeStr"
             else -> "belirtilen zamanda"
         }
     }
